Refactor contextual dictionaries.
